Add end-to-end C++ exception handling (try/catch/throw) for SLOW-32
Full DWARF CFI-based exception handling across all toolchain layers:

LLVM backend: DwarfCFI exception type, DwarfRegNum annotations for all 32
registers, CFI prologue/epilogue emission, FRAMEADDR/RETURNADDR lowering,
exception pointer in R3 and selector in R4.

Assembler: Parse .cfi_* directives and emit binary DWARF .eh_frame CIE/FDE
records. Support .gcc_except_table as pass-through section. Fix critical bug
where duplicate .uleb128 handler (simple parse_immediate) shadowed the
label-difference-aware handler, zeroing the entire LSDA call site table.

Linker: Merge .eh_frame sections from multiple objects (skip intermediate
zero terminators), inject __eh_frame_start/__eh_frame_end discovery symbols.

Runtime: DWARF stack unwinder (unwind.c) with .eh_frame parser, FDE lookup,
CFA interpreter, two-phase unwinding. Assembly glue (unwind_asm.s) for
register save/restore. C++ ABI layer: __cxa_allocate_exception, __cxa_throw,
__cxa_begin_catch, __cxa_end_catch, __gxx_personality_v0 with LSDA parsing
and type matching.

Key bugs fixed during bring-up: .eh_frame zero terminator parsing (break
to continue), MAX_CIES too small for merged sections (8 to 64), missing
ip-1 DWARF convention for FDE lookup, step_frame setting personality/LSDA
from wrong frame (added lookup_frame_info after step_frame).

All 60 regression tests pass including new cpp-exception-basic test.

+++ llvm-backend/SLOW32/SLOW32ISelDAGToDAG.cpp
@@ -90,7 +90,7 @@ public:
         int64_t CVal = CN->getSExtValue();
         if (isInt<12>(CVal)) {
           Base = N.getOperand(0);
-          Offset = CurDAG->getTargetConstant(CVal, DL, MVT::i32);
+          Offset = CurDAG->getTargetConstant(static_cast<uint32_t>(CVal), DL, MVT::i32);
           return true;
         }
       }
@@ -106,7 +106,7 @@ public:
         int64_t CVal = CN->getSExtValue();
         if (isInt<12>(CVal)) {
           Base = N.getOperand(1);
-          Offset = CurDAG->getTargetConstant(CVal, DL, MVT::i32);
+          Offset = CurDAG->getTargetConstant(static_cast<uint32_t>(CVal), DL, MVT::i32);
           return true;
         }
       }
@@ -141,7 +141,7 @@ public:
         if (auto *C = dyn_cast<ConstantSDNode>(Op1)) {
           int64_t Imm = C->getSExtValue();
           if (isInt<12>(Imm)) {
-            SDValue ImmOp = CurDAG->getTargetConstant(Imm, DL, MVT::i32);
+            SDValue ImmOp = CurDAG->getTargetConstant(static_cast<uint32_t>(Imm), DL, MVT::i32);
             SDNode *Res =
                 CurDAG->getMachineNode(SLOW32::ADDI, DL, N->getValueType(0),
                                        Op0, ImmOp);
@@ -151,7 +151,7 @@ public:
         } else if (auto *C = dyn_cast<ConstantSDNode>(Op0)) {
           int64_t Imm = C->getSExtValue();
           if (isInt<12>(Imm)) {
-            SDValue ImmOp = CurDAG->getTargetConstant(Imm, DL, MVT::i32);
+            SDValue ImmOp = CurDAG->getTargetConstant(static_cast<uint32_t>(Imm), DL, MVT::i32);
             SDNode *Res =
                 CurDAG->getMachineNode(SLOW32::ADDI, DL, N->getValueType(0),
                                        Op1, ImmOp);
